wip: refactor external component as polymer component
